requirements.txt
# Core
Django==5.0.1
django-ninja==1.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.2.1
django-cors-headers==4.3.1

# Database
redis==5.2.1

# Validation
pydantic==2.10.6
pydantic-settings==2.7.1

# Utilities
python-slugify==8.0.4
pillow>=10.3.0

# Development
ipython==8.31.0
